+++ src/fixed_point/combinational/harnesses/complex_multiplier.py
@@ -1,9 +1,3 @@
-# -*- coding: utf-8 -*-
-# @Author: UnsignedByte
-# @Date:   2022-11-12 18:33:26
-# @Last Modified by:   UnsignedByte
-# @Last Modified time: 2022-11-12 18:33:48
-
 from pymtl3 import *
 from pymtl3.stdlib import stream
 from pymtl3.passes.backends.verilog import *
